+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * :feature:`59` Fail AMQP client command when the consumer is not connected.
 * :support:`-` Typing: `.BaseCommand` now accepts a generic with the class of the actor.
 * :feature:`61` `.Model` and `.TronModel` callbacks receive the model and the updated keyword again. This is done in a backwards compatible manner; if the callback has a single argument in its signature it will receive only the model.
+* :bug:`-` Deal with a case in which the loop for a ``CallbackMixIn`` may not be running at the time at which the callback needs to be invoked.
 
 * :release:`0.7.1 <2021-02-21>`
 * :feature:`-` Add the option to update the object of the parser object by setting `.ClickParser.context_obj`.